refactor: move CLI run model policy to runtime
