+++ src/test-kernel/cli/MultiAgentRunCommand.vitest.mts
@@ -105,7 +105,7 @@ vi.mock('@cli/runtime/multiAgentPresets', () => {
   };
 });
 
-vi.mock('@cli/commands/_helpers/modelArg', () => ({
+vi.mock('@cli/runtime/runModel', () => ({
   buildHeadlessRunContext: vi.fn((context: CliContext, model: string) => ({
     ...context,
     helperModel: model,
